timer: track loop time in nanoseconds
The loop clock is tracked in milliseconds. This means that the difference
between 2 successive updates to the loop clock could show a millisecond
difference when only a nanosecond has passed. Timers are triggered in
this way. In NodeJS, this means that setTimeout used with 1 millisecond
could be triggered immediately.

This change only changes the internal loop clock to nanosecond resolution
but all external api's will still use millisecond resolution.